# chore: redo gauss calculations