Replace https://github.com/chobie/go-gaussian with custom implementation.

Changes:
 - go-gausian uses a custom `Erfc` function, that is slower than the one in the `math`. The custom implementatation calculates an approximation of the complementary error function using a polinomial. https://github.com/chobie/go-gaussian/blob/53c09d90eeaff40b27cd89352937d02d57dac506/gaussian.go#L43-L55
 - Use constants for sqrt of 2 and sqrt of 2*Pi
 - Avoid math.Pow(x, 2) and just multiply
 - Avoid math.Sqrt by providing the standard deviation instead of the variance
 - Carefullly document

```
→ benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: github.com/form3tech-oss/f1/v2/internal/trigger/gaussian
                                  │     old.txt     │             new.txt              │
                                  │     sec/op      │     sec/op       vs base         │
_calculateVolume/1400TPS-10         0.0018220n ± 0%   0.0006767n ± 0%  -62.86% (n=100)
_calculateVolume/50TPS-10           0.0018215n ± 0%   0.0006759n ± 0%  -62.89% (n=100)
_calculateVolume/10TPS-10           0.0018225n ± 0%   0.0006770n ± 0%  -62.85% (n=100)
_calculateVolume/10TPS_2-10         0.0018215n ± 0%   0.0006773n ± 0%  -62.81% (n=100)
_calculateVolume/10TPS_no_unit-10   0.0018150n ± 0%   0.0006685n ± 0%  -63.17% (n=100)
_calculateVolume/1TPms-10           0.0018230n ± 0%   0.0006772n ± 0%  -62.85% (n=100)
_calculateVolume/1000TPms-10        0.0018220n ± 0%   0.0006766n ± 0%  -62.86% (n=100)
geomean                              0.001821n        0.0006756n       -62.90%
```